+++ ecd-20201002/alibabacloud_ecd20201002/client.py
@@ -1497,6 +1497,8 @@ class Client(OpenApiClient):
             query['SessionId'] = request.session_id
         if not UtilClient.is_unset(request.session_token):
             query['SessionToken'] = request.session_token
+        if not UtilClient.is_unset(request.uuid):
+            query['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -1541,6 +1543,8 @@ class Client(OpenApiClient):
             query['SessionId'] = request.session_id
         if not UtilClient.is_unset(request.session_token):
             query['SessionToken'] = request.session_token
+        if not UtilClient.is_unset(request.uuid):
+            query['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -2324,7 +2328,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> ecd_20201002_models.StartDesktopsResponse:
         """
-        The ID of the client.
+        The cloud computers that you want to start must be in the Stopped state. After you call this operation, the cloud computers enter the Running state.
         
         @param request: StartDesktopsRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2348,6 +2352,8 @@ class Client(OpenApiClient):
             query['RegionId'] = request.region_id
         if not UtilClient.is_unset(request.session_id):
             query['SessionId'] = request.session_id
+        if not UtilClient.is_unset(request.uuid):
+            query['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -2373,7 +2379,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> ecd_20201002_models.StartDesktopsResponse:
         """
-        The ID of the client.
+        The cloud computers that you want to start must be in the Stopped state. After you call this operation, the cloud computers enter the Running state.
         
         @param request: StartDesktopsRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2397,6 +2403,8 @@ class Client(OpenApiClient):
             query['RegionId'] = request.region_id
         if not UtilClient.is_unset(request.session_id):
             query['SessionId'] = request.session_id
+        if not UtilClient.is_unset(request.uuid):
+            query['Uuid'] = request.uuid
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -2421,7 +2429,7 @@ class Client(OpenApiClient):
         request: ecd_20201002_models.StartDesktopsRequest,
     ) -> ecd_20201002_models.StartDesktopsResponse:
         """
-        The ID of the client.
+        The cloud computers that you want to start must be in the Stopped state. After you call this operation, the cloud computers enter the Running state.
         
         @param request: StartDesktopsRequest
         @return: StartDesktopsResponse
@@ -2434,7 +2442,7 @@ class Client(OpenApiClient):
         request: ecd_20201002_models.StartDesktopsRequest,
     ) -> ecd_20201002_models.StartDesktopsResponse:
         """
-        The ID of the client.
+        The cloud computers that you want to start must be in the Stopped state. After you call this operation, the cloud computers enter the Running state.
         
         @param request: StartDesktopsRequest
         @return: StartDesktopsResponse
@@ -2546,7 +2554,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> ecd_20201002_models.StopDesktopsResponse:
         """
-        The cloud desktops that you want to stop by calling this operation must be in the Running state. If the call is successful, the cloud desktops enter the Stopped state.
+        The cloud computers that you want to stop must be in the Running state. After you call this operation, the cloud computers enter the Stopped state.
         
         @param request: StopDesktopsRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2597,7 +2605,7 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> ecd_20201002_models.StopDesktopsResponse:
         """
-        The cloud desktops that you want to stop by calling this operation must be in the Running state. If the call is successful, the cloud desktops enter the Stopped state.
+        The cloud computers that you want to stop must be in the Running state. After you call this operation, the cloud computers enter the Stopped state.
         
         @param request: StopDesktopsRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -2647,7 +2655,7 @@ class Client(OpenApiClient):
         request: ecd_20201002_models.StopDesktopsRequest,
     ) -> ecd_20201002_models.StopDesktopsResponse:
         """
-        The cloud desktops that you want to stop by calling this operation must be in the Running state. If the call is successful, the cloud desktops enter the Stopped state.
+        The cloud computers that you want to stop must be in the Running state. After you call this operation, the cloud computers enter the Stopped state.
         
         @param request: StopDesktopsRequest
         @return: StopDesktopsResponse
@@ -2660,7 +2668,7 @@ class Client(OpenApiClient):
         request: ecd_20201002_models.StopDesktopsRequest,
     ) -> ecd_20201002_models.StopDesktopsResponse:
         """
-        The cloud desktops that you want to stop by calling this operation must be in the Running state. If the call is successful, the cloud desktops enter the Stopped state.
+        The cloud computers that you want to stop must be in the Running state. After you call this operation, the cloud computers enter the Stopped state.
         
         @param request: StopDesktopsRequest
         @return: StopDesktopsResponse