+++ ecd-20201002/ChangeLog.md
@@ -1,3 +1,9 @@
+2024-04-25 Version: 1.1.1
+- Update API GetLoginToken: update response param.
+- Update API RebootDesktops: add param Uuid.
+- Update API StartDesktops: add param Uuid.
+
+
 2024-03-21 Version: 1.1.0
 - Support API GetConnectionTicket.
 - Support API QueryEdsAgentReportConfig.